requirements.txt
openai
pydantic
tavily-python
spacy
pandas
langchain
langchain-core
nltk
tiktoken
google
boto3
